chore: refactor file reader and writer
